Refactor credential retrieval
