+++ provider/pkg/googleclient/credentials.go
@@ -1,0 +1,161 @@
+package googleclient
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/golang/glog"
+	"github.com/pkg/errors"
+	"golang.org/x/oauth2"
+	"golang.org/x/oauth2/google"
+	"google.golang.org/api/option"
+	"google.golang.org/api/transport"
+)
+
+type tokenSourceFunc func() (*oauth2.Token, error)
+
+func (t tokenSourceFunc) Token() (*oauth2.Token, error) {
+	return t()
+}
+
+func validatedTokenSource(src oauth2.TokenSource) oauth2.TokenSource {
+	return tokenSourceFunc(func() (*oauth2.Token, error) {
+		t, err := src.Token()
+		if err != nil {
+			return nil, err
+		}
+
+		if !t.Valid() {
+			return nil, fmt.Errorf("expired token provided")
+		}
+
+		return t, nil
+	})
+}
+
+func newReuseValidatedTokenSource(t *oauth2.Token, src oauth2.TokenSource) *reuseValidatedTokenSource {
+	return &reuseValidatedTokenSource{
+		src: validatedTokenSource(oauth2.ReuseTokenSource(t, src)),
+	}
+}
+
+type reuseValidatedTokenSource struct {
+	src oauth2.TokenSource
+}
+
+func (r *reuseValidatedTokenSource) Token() (*oauth2.Token, error) {
+	return r.src.Token()
+}
+
+type credentialValidator interface {
+	isValid() (bool, error)
+}
+
+type credentialValidatorFunc func() (bool, error)
+
+func (c credentialValidatorFunc) isValid() (bool, error) {
+	return c()
+}
+
+type credentialRetriever interface {
+	getCredentials(ctx context.Context, c Config) (*google.Credentials, error)
+}
+
+type getCredentialsFunc func(context.Context, Config) (*google.Credentials, error)
+
+func (g getCredentialsFunc) getCredentials(ctx context.Context, c Config) (*google.Credentials, error) {
+	return g(ctx, c)
+}
+
+func defaultCredentialRetriever() credentialRetriever {
+	return getCredentialsFunc(getCredentials)
+}
+
+func getCredentials(ctx context.Context, c Config) (*google.Credentials, error) {
+	if c.AccessToken != "" {
+		contents, _, err := pathOrContents(c.AccessToken)
+		if err != nil {
+			return nil, errors.Wrapf(err, "Error loading access token")
+		}
+		token := &oauth2.Token{AccessToken: contents}
+
+		if c.ImpersonateServiceAccount != "" {
+			opts := []option.ClientOption{
+				option.WithTokenSource(oauth2.StaticTokenSource(token)),
+				option.ImpersonateCredentials(c.ImpersonateServiceAccount, c.ImpersonateServiceAccountDelegates...),
+				option.WithScopes(c.Scopes...),
+			}
+			creds, err := transport.Creds(ctx, opts...)
+			if err != nil {
+				return nil, err
+			}
+			return creds, nil
+		}
+
+		glog.V(9).Info("Authenticating using configured Google JSON Access Token...")
+		glog.V(9).Infof("  -- Scopes: %s", c.Scopes)
+
+		return &google.Credentials{
+			TokenSource: staticTokenSource{oauth2.StaticTokenSource(token)},
+		}, nil
+	}
+
+	if c.Credentials != "" {
+		contents, _, err := pathOrContents(c.Credentials)
+		if err != nil {
+			return nil, errors.Wrapf(err, "error loading credentials")
+		}
+		if c.ImpersonateServiceAccount != "" {
+			opts := []option.ClientOption{
+				option.WithCredentialsJSON([]byte(contents)),
+				option.ImpersonateCredentials(c.ImpersonateServiceAccount, c.ImpersonateServiceAccountDelegates...),
+				option.WithScopes(c.Scopes...),
+			}
+			creds, err := transport.Creds(ctx, opts...)
+			if err != nil {
+				return nil, err
+			}
+
+			glog.V(9).Info("Authenticating using configured Google JSON Credentials and Impersonate Service Account...")
+			glog.V(9).Infof("   -- Scopes: %s", c.Scopes)
+			return creds, nil
+		}
+		creds, err := google.CredentialsFromJSON(ctx, []byte(contents), c.Scopes...)
+		if err != nil {
+			return nil, errors.Wrapf(err, "unable to parse credentials from %q", contents)
+		}
+
+		glog.V(9).Info("Authenticating using configured Google JSON Credentials...")
+		glog.V(9).Infof("   -- Scopes: %s", c.Scopes)
+		return creds, nil
+	}
+
+	if c.ImpersonateServiceAccount != "" {
+		opts := option.ImpersonateCredentials(c.ImpersonateServiceAccount, c.ImpersonateServiceAccountDelegates...)
+		creds, err := transport.Creds(ctx, opts, option.WithScopes(c.Scopes...))
+		if err != nil {
+			return nil, err
+		}
+
+		glog.V(9).Info("Authenticating using configured Impersonate Service Account...")
+		glog.V(9).Infof("   -- Scopes: %s", c.Scopes)
+		return creds, nil
+	}
+
+	glog.V(9).Info("Authenticating using DefaultClient...")
+	glog.V(9).Infof("   -- Scopes: %s", c.Scopes)
+
+	defaultTS, err := google.DefaultTokenSource(context.Background(), c.Scopes...)
+	if err != nil {
+		return nil, fmt.Errorf("Attempted to load application default credentials since neither `credentials` nor `access_token` was set in the provider block.  No credentials loaded. To use your gcloud credentials, run 'gcloud auth application-default login'.  Original error: %w", err)
+	}
+	return &google.Credentials{
+		TokenSource: defaultTS,
+	}, err
+}
+
+// TODO: Why do we need this?
+// staticTokenSource is used to be able to identify static token sources without reflection.
+type staticTokenSource struct {
+	oauth2.TokenSource
+}
